requirements.txt
astropy==7.0.0
ccdproc==2.4.2
ipykernel==6.19.2
matplotlib==3.10.0
numba==0.57.0
numpy==1.24.4
psutil==5.9.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.67.1
watchdog==6.0.0

# cupy
cupy-cuda12x

# eclaire
git+https://github.com/7DimensionalTelescope/Eclaire
